Fixing aws_cloudwatch_log_retention tests (for real)
I reverted the tests file to how it was before and properly implemented a mock of the gql api (now that I kinda understand what is going on).
From there I replaced the instances in the test file of accounts stored as dicts to AWSAccountV1s.
I had to add some code and change some code to make it compatible with parts of the code base that expects a dict.