Typed queries for aws-cloudwatch-log-retention integration (#4952)
* Typed queries for aws-cloudwatch-log-retention integration

Added custom gql definition for `aws-cloudwatch-log-retention` to replace the non-typed `get_aws_accounts` query from `queries.py`. The only option necessary for this integration is `cleanup`, and so that is what the custom gql definition includes. The query is defined in `reconcile/typed_queries/aws_cloudwatch_log_retention/aws_accounts.py`. 

Modified the the `integration.py` to use the new query, as well as convert instances and typing of dict to the appropriate types.

Tests were modified to be compatible with the new typed system.
I implemented a mock of the gql api, and changed dict indexing to property accessing using a dot operator.

Streamlined conditionals.
Made sure everything stayed as typed as it could.

* Fixing typing issue.

Removed `AWSAccountCleanupOptionAMIV1` from the gql query.
Modified filter to check if the cleanup option is of type `AWSAccountCleanupOptionCloudWatchV1` in `get_desired_cleanup_options_by_region` in `integration.py`
Changed to access attributes of typed class rather than access as a dict