Use debug-container for sql queries (#5062)
This is to replace sqlQuery.imageRepo which is the base to use mysql and
and postgres client commands.

In sql-query and and database-access-manager we use the
`quay.io/app-sre-/postgres|mysql:<major version>` mirrored containers to run
the queries, but we can't use those mirrors any longer for Konflux reasons...
The problem is that we can't replace them with just the postgres|mysql
registry version because the cadence of release in registry is not fast enough,
e.g.  they're currently in postgres version 16 whereas AWS is already
offering 18 in beta. So the proposal here is to use `debug-container:latest`.
That gives us a good chance of using something that we keep up to date.

xref: https://github.com/app-sre/qontract-schemas/pull/872

APPSRE-12189

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>